FUnctional decomposition and code review complete
